use new StorageOperator class from vsc-filesystems
